Change the converter for getTankInfo results from IFluidTank to FluidTankInfo
